Move visual attachment code to the separate folder
